requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
atlassian-python-api==3.41.14
attrs==24.2.0
beautifulsoup4==4.12.3
black==24.8.0
bs4==0.0.2
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
commonmark==0.9.1
crashtest==0.4.1
databricks-sdk==0.32.0
databricks-sql-connector==3.4.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dictdiffer==0.9.0
dirtyjson==1.0.8
distro==1.9.0
et-xmlfile==1.1.0
faiss-cpu==1.8.0.post1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.66.1
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
langchain-text-splitters==0.2.4
langsmith==0.1.108
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
markdownify==0.13.1
marshmallow==3.22.0
md2notion==2.4.1
mdurl==0.1.2
mdutils==1.6.0
mistletoe==1.4.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
notion==0.0.28
notion-client==2.2.1
notion2md==2.9.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.44.0
openpyxl==3.1.5
orjson==3.10.7
pandas==2.2.2
pathspec==0.12.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
pypdf==4.3.1
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
rank-bm25==0.2.2
rapidfuzz==3.9.6
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
setuptools==72.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
striprtf==0.0.26
tenacity==8.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
thrift==0.20.0
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.11
google-cloud-storage==2.18.2